requirements.txt
tensorflow==2.15.0
onnx-tf==1.10.0
tensorflow-probability==0.23.0
torch
pandas
pillow
torchvision
onnx
onnxruntime
numpy